First collision working

+++ examples/game/components/body_component.h
@@ -2,8 +2,8 @@
 
 struct BodyComponent
 {
-  BodyComponent(float width, float height, float offsetx, float offsety, bool isStatic) : Width(width), Height(height), OffsetX(offsetx), OffsetY(offsety), Static(isStatic){}
-  float Width, Height, OffsetX, OffsetY;
+  BodyComponent(int width, int height, int offsetx, int offsety, bool isStatic) : Width(width), Height(height), OffsetX(offsetx), OffsetY(offsety), Static(isStatic){}
+  int Width, Height, OffsetX, OffsetY;
   bool Static;
 };
